chore: move concat into separate file
