+++ src/operators/index.ts
@@ -11,10 +11,11 @@ export { combineAll } from '../internal/operators/combineAll';
 export { combineLatestAll } from '../internal/operators/combineLatestAll';
 export { combineLatest } from '../internal/operators/combineLatest';
 export { combineLatestWith } from '../internal/operators/combineLatestWith';
+export { concat } from '../internal/operators/concat';
 export { concatAll } from '../internal/operators/concatAll';
 export { concatMap } from '../internal/operators/concatMap';
 export { concatMapTo } from '../internal/operators/concatMapTo';
-export { concat, concatWith } from '../internal/operators/concatWith';
+export { concatWith } from '../internal/operators/concatWith';
 export { connect } from '../internal/operators/connect';
 export { count } from '../internal/operators/count';
 export { debounce } from '../internal/operators/debounce';
